tests/wpt/tests/intersection-observer/padding-clip.html

<!doctype html>
<meta charset="utf-8">
<meta name="viewport" content="width=device-width,initial-scale=1">
<title>Scrollport is used rather than content rect to compute intersection ratio</title>
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://github.com/w3c/IntersectionObserver/issues/504">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#root {
  overflow: hidden;
  position: relative;
  box-sizing: border-box;
  width: 100px;
  height: 100px;
  padding: 10px;
}
#target {
  width: 100px;
  height: 100px;
  background-color: green;
  /* Shift into the padding area */
  position: relative;
  left: -10px;
  top: -10px;
}
</style>
<div id="root">
  <div id="target"></div>
</div>
<script>
promise_test(async function(t) {
  const target = document.getElementById("target");
  assert_true(!!target, "target exists");
  const root = document.getElementById("root");
  assert_true(!!root, "root exists");
  const entries = await new Promise(resolve => {
    const observer = new IntersectionObserver(resolve, { root });
    observer.observe(target);
  });
  assert_equals(entries.length, 1, "Should have one entry.");
  assert_equals(entries[0].intersectionRatio, 1, "Should be completely visible.");
});
</script>
