third_party/blink/web_tests/platform/mac-mac12-arm64/virtual/disable-intersection-optimization/external/wpt/intersection-observer/padding-clip-expected.txt

This is a testharness.js-based test.
[FAIL] Scrollport is used rather than content rect to compute intersection ratio
  assert_equals: Should be completely visible. expected 1 but got 0.6399999856948853
Harness: the test ran to completion.
